api/requirements.txt
absl-py==0.4.1
astor==0.7.1
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.49.0
boto3==1.9.164
botocore==1.12.164
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cvxpy==1.0.23
cycler==0.10.0
defusedxml==0.5.0
dill==0.2.9
Django==2.2.10
django-allauth==0.41.0
django-filter==2.0.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
drf-yasg==1.11.0
dtcwt==0.12.0
ecos==2.0.7.post1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
future==0.17.1
gast==0.2.0
gensim==3.7.3
googledrivedownloader==0.4
grpcio==1.14.2
html5lib==0.9999999
idna==2.7
imageio==2.3.0
imutils==0.5.2
inflection==0.3.1
itsdangerous==0.24
itypes==1.1.0
Jinja2>=2.10.1
jmespath==0.9.4
JPype1==0.6.3
kiwisolver==1.0.1
konlp==0.0.8
konlpy==0.5.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mglearn==0.1.7
mido==1.2.8
mpmath==1.1.0
multiprocess==0.70.7
numpy==1.16.3
oauthlib==3.0.1
opencv-python==4.0.0.21
osqp==0.5.0
pandas==0.24.2
Pillow==6.2.0
pretty-midi==0.2.8
progress==1.3
protobuf==3.6.1
psycopg2==2.7.6
PuLP==1.6.10
PyJWT==1.7.1
pyparsing==2.2.0
pypianoroll==0.4.6
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyWavelets==1.0.1
PyYAML>=4.2b1
redis==3.3.11
requests>=2.20.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.75
s3transfer==0.2.1
scikit-learn==0.20.3
scipy==1.1.0
scs==2.1.0
seaborn==0.9.0
six==1.11.0
smart-open==1.8.4
social-auth-app-django==3.1.0
social-auth-core==3.0.0
SQLAlchemy>=1.3.0
sympy==1.4
tensorboard==1.10.0
termcolor==1.1.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.24.2
Werkzeug==0.15.3
